+++ custom_components/nodered/discovery.py
@@ -2,11 +2,11 @@
 import asyncio
 import logging
 
+from homeassistant.core import HomeAssistant
 from homeassistant.helpers.dispatcher import (
     async_dispatcher_connect,
     async_dispatcher_send,
 )
-from homeassistant.helpers.typing import HomeAssistantType
 
 from .const import (
     CONF_BINARY_SENSOR,
@@ -48,9 +48,7 @@ CONFIG_ENTRY_IS_SETUP = "config_entry_is_setup"
 DISCOVERY_DISPATCHED = "discovery_dispatched"
 
 
-async def start_discovery(
-    hass: HomeAssistantType, hass_config, config_entry=None
-) -> bool:
+async def start_discovery(hass: HomeAssistant, hass_config, config_entry=None) -> bool:
     """Initialize of Node-RED Discovery."""
 
     async def async_device_message_received(msg, connection):
@@ -113,6 +111,6 @@ async def start_discovery(
     )
 
 
-def stop_discovery(hass: HomeAssistantType):
+def stop_discovery(hass: HomeAssistant):
     """Remove discovery dispatcher."""
     hass.data[DOMAIN_DATA][DISCOVERY_DISPATCHED]()